% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwrmodel.R
\name{fwrmodel}
\alias{fwrmodel}
\title{Finlay-Wilkinson Regression Model}
\usage{
fwrmodel(data, env_cols, genotype_col, yield_cols)
}
\arguments{
\item{data}{A data frame containing the yield data.}

\item{env_cols}{A character vector of column names that describe the environment (e.g., Fertilizer, Irrigation).}

\item{genotype_col}{The name of the column containing the genotype or cultivar names.}

\item{yield_col}{The name of the column containing the yield values.}
}
\value{
A data frame with the slopes for each genotype.
}
\description{
This function calculates the slope of yield against the environment index for each genotype.
}
